refactor(path): update import module name
