+++ packages/cli/src/commands/migrate.ts
@@ -2,7 +2,7 @@ import { execa, ExecaError } from 'execa';
 import path from 'path';
 import yoctoSpinner from 'yocto-spinner';
 
-import { getEnginePath } from '../utils.js';
+import { getEnginePath } from '../lib.js';
 
 const spinner = yoctoSpinner({ text: 'Migrating Database\n' });
