+++ packages/cli/src/commands/serve.ts
@@ -7,7 +7,7 @@ import path from 'path';
 import fsExtra from 'fs-extra/esm';
 import fs from 'fs/promises';
 
-import { getFirstExistingFile } from '../utils.js';
+import { getFirstExistingFile } from '../lib.js';
 import { bundle } from '../utils/bundle.js';
 
 import { EXPRESS_SERVER } from './deploy/server.js';
